Struct nng::ListenerBuilder

pub struct ListenerBuilder { /* fields omitted */ }

Configuration utility for nanomsg-next-generation listeners.

This object allows for the configuration of listeners before they are started. If it is not necessary to change listener settings or to close the listener without closing the socket, then Socket::listen provides a simpler interface.

Implementations

impl ListenerBuilder

pub fn new(socket: &Socket, url: &str) -> Result<Self>

Creates a new Listener object associated with the given socket.

Note that this does not start the Listener In order to start the listener, this object must be consumed by ListenerBuilder::start.

Errors

AddressInvalid: An invalid url was specified.

Closed: The socket is not open.

OutOfMemory: Insufficient memory.

pub fn start(self) -> Result<Listener, (Self, Error)>

Cause the Listener to start listening on the address with which it was created.

The returned handle controls the life of the Listener. If it is dropped, the Listener is shut down and no more messages will be received on it.

Errors

Closed: The socket is not open.